rename pages->views, layouts -> panels
