+++ src/panels/Inspector/CourseDetails.tsx
