+++ src/panels/Inspector/Inspector.tsx
