+++ src/panels/Inspector/index.ts
