+++ src/views/CourseMap/CourseMap.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 
 import CourseGraph from './components/CourseGraph';
-import CourseDetails from 'layouts/Inspector/CourseDetails';
-import LeftPanel from 'layouts/DarsPanel';
+import LeftPanel from 'panels/DarsPanel';
+import CourseDetails from 'panels/Inspector/CourseDetails';
 import { Course } from 'types/Course';
 import './style.scss';
