+++ src/views/CourseMap/index.ts
