+++ src/views/CourseMap/style.scss
